requirements.txt
openai==1.57.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
faiss-cpu== 1.9.0.post1
requests==2.32.3
ratelimit==2.2.1
nltk==3.8.1
transformers==4.36.0
torch==2.5.1
scikit-learn==1.3.2
tqdm==4.66.1